+++ CHANGELOG.md
@@ -2,9 +2,7 @@
 
 ## Added
 
-## Fixed
-
-## Changed
+- Add MapProvider, DerefMapProvider, and integration with lambdaisland.cli
 
 # 0.1.6 (2024-12-17 / 0e8e619)